lots more

+++ api/Oc6.Bold/Controllers/PlayerController.cs
@@ -17,10 +17,12 @@ namespace Oc6.Bold.Controllers
     public class PlayerController : ControllerBase
     {
         private readonly PlayerService playerService;
+        private readonly UserService userService;
 
-        public PlayerController(PlayerService playerService)
+        public PlayerController(PlayerService playerService, UserService userService)
         {
             this.playerService = playerService;
+            this.userService = userService;
         }
 
         [HttpGet("{id}")]
@@ -33,30 +35,35 @@ namespace Oc6.Bold.Controllers
 
         [HttpGet("self")]
         public async Task<PlayerDto> Self() =>
-            await playerService.GetOrCreateSelf(
-                User.NameFromClaims(),
-                User.EmailFromClaims(),
-                User.IsAdmin());
+            await playerService.GetOrCreateSelf();
 
-        [HttpPost("self")]
+        [HttpPut("self")]
         public async Task<PlayerDto> Self([FromBody] PlayerDto player) =>
-            await playerService.UpdateOrCreateSelf(
-                User.NameFromClaims(),
-                User.EmailFromClaims(),
-                player.Name,
-                User.IsAdmin());
+            await playerService.UpdateSelf(player.Name);
 
         [HttpPost]
         [AdminPolicyAuthorize]
         public async Task<PlayerDto> Post([FromBody] PlayerDto dto) =>
-            await playerService.Create(dto.Email, dto.Auth0UserId, User.IsAdmin());
+            await playerService.Create(dto.Email);
 
         [HttpPut]
         public async Task<PlayerDto> Put([FromBody] PlayerDto dto) =>
             await playerService.Update(dto.Id, dto.Name);
 
         [HttpDelete("{id}")]
-        public async Task Delete([FromRoute] int id) =>
+        [ProducesDefaultResponseType(typeof(void))]
+        public async Task<IActionResult> Delete([FromRoute] int id)
+        {
+            var userId = await userService.GetCurrentUserIdAsync();
+
+            if (userId == id)
+            {
+                return BadRequest();
+            }
+
             await playerService.Delete(id);
+
+            return Ok();
+        }
     }
 }
